Move ert_plugin, ert_script and external_ert_script to ert.plugins
