+++ docs/ert/conf.py
@@ -178,7 +178,7 @@ intersphinx_mapping = {"python": ("https://docs.python.org/3", None)}
 
 # Ignore unresolved references
 nitpick_ignore = [
-    ("py:class", "ert.config.ert_script.ErtScript"),
+    ("py:class", "ert.plugins.ert_script.ErtScript"),
     ("py:class", "Ensemble"),
     ("py:class", "ESSettings"),
     ("py:class", "Storage"),
