+++ src/ert/gui/tools/plugins/plugin.py
@@ -3,8 +3,8 @@ from __future__ import annotations
 import inspect
 from typing import TYPE_CHECKING, Any
 
-from ert.config import ErtPlugin
 from ert.config.workflow_fixtures import WorkflowFixtures
+from ert.plugins import ErtPlugin
 
 if TYPE_CHECKING:
     from PyQt6.QtWidgets import QWidget
