+++ src/ert/plugins/hook_implementations/workflows/export_misfit_data.py
@@ -4,7 +4,7 @@ from typing import TYPE_CHECKING, Any
 
 import pandas as pd
 
-from ert.config.ert_script import ErtScript
+from ert import ErtScript
 from ert.exceptions import StorageError
 
 if TYPE_CHECKING:
