+++ src/ert/workflow_runner.py
@@ -5,8 +5,9 @@ from concurrent import futures
 from concurrent.futures import Future
 from typing import Any, Self
 
-from ert.config import ErtScript, ExternalErtScript, Workflow, WorkflowJob
+from ert.config import Workflow, WorkflowJob
 from ert.config.workflow_fixtures import WorkflowFixtures
+from ert.plugins import ErtScript, ExternalErtScript
 
 
 class WorkflowJobRunner:
