+++ tests/ert/unit_tests/config/test_ert_config.py
@@ -312,7 +312,7 @@ def test_that_get_plugin_jobs_fetches_exactly_ert_plugins():
     Path("plugin.py").write_text(
         dedent(
             """
-            from ert.config import ErtPlugin
+            from ert.plugins import ErtPlugin
 
             class Plugin(ErtPlugin):
                 def run(self, *args):
